Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -3,10 +3,10 @@
 char* welcome() {
     char ch;
     scanf("%c",&ch);
-    return (int)ch;
+    return ch;
 }
 
 int main() {
-    printf("ASCII Value: %d", welcome());
+    printf("ASCII Value: %d", (int)welcome());
     return 0;
 }
